Pulled out escalation fetch logic to index file
